back/requirements.txt
# Core framework
Flask==3.0.3
flask-cors==4.0.0
python-dotenv==1.0.1
requests==2.32.3

# OpenAI + embeddings
openai==1.40.0
qdrant-client==1.9.2  # only if you use SDK; safe to keep

# Production servers
gunicorn==21.2.0
waitress==3.0.0

# Document processing
python-docx==1.1.2
pdfplumber==0.11.4
openpyxl==3.1.5
beautifulsoup4==4.12.3
lxml==5.3.0
chardet==5.2.0
httpx==0.27.2
httpcore==1.0.5


